update_upm_to_1.9.4

+++ Editor/AssetBundleTree.cs
@@ -4,12 +4,12 @@ using System.Collections.Generic;
 using UnityEditor.IMGUI.Controls;
 using System.Linq;
 using System;
-
+using System.IO;
 
 namespace AssetBundleBrowser
 {
     internal class AssetBundleTree : TreeView
-    { 
+    {
         AssetBundleManageTab m_Controller;
         private bool m_ContextOnItem = false;
         List<UnityEngine.Object> m_EmptyObjectList = new List<UnityEngine.Object>();
@@ -52,17 +52,17 @@ namespace AssetBundleBrowser
             GUI.color = old;
 
             var message = bundleItem.BundleMessage();
-            if(message.severity != MessageType.None)
+            if (message.severity != MessageType.None)
             {
                 var size = args.rowRect.height;
                 var right = args.rowRect.xMax;
                 Rect messageRect = new Rect(right - size, args.rowRect.yMin, size, size);
-                GUI.Label(messageRect, new GUIContent(message.icon, message.message ));
+                GUI.Label(messageRect, new GUIContent(message.icon, message.message));
             }
         }
 
         protected override void RenameEnded(RenameEndedArgs args)
-        { 
+        {
             base.RenameEnded(args);
             if (args.newName.Length > 0 && args.newName != args.originalName)
             {
@@ -95,7 +95,7 @@ namespace AssetBundleBrowser
                 foreach (var id in selectedIds)
                 {
                     var item = FindItem(id, rootItem) as AssetBundleModel.BundleTreeItem;
-                    if(item != null && item.bundle != null)
+                    if (item != null && item.bundle != null)
                     {
                         item.bundle.RefreshAssetList();
                         selectedBundles.Add(item.bundle);
@@ -109,7 +109,7 @@ namespace AssetBundleBrowser
         public override void OnGUI(Rect rect)
         {
             base.OnGUI(rect);
-            if(Event.current.type == EventType.MouseDown && Event.current.button == 0 && rect.Contains(Event.current.mousePosition))
+            if (Event.current.type == EventType.MouseDown && Event.current.button == 0 && rect.Contains(Event.current.mousePosition))
             {
                 SetSelection(new int[0], TreeViewSelectionOptions.FireSelectionChanged);
             }
@@ -127,13 +127,18 @@ namespace AssetBundleBrowser
             List<AssetBundleModel.BundleTreeItem> selectedNodes = new List<AssetBundleModel.BundleTreeItem>();
             GenericMenu menu = new GenericMenu();
 
-            if (!AssetBundleModel.Model.DataSource.IsReadOnly ()) {
-                menu.AddItem(new GUIContent("Add new bundle"), false, CreateNewBundle, selectedNodes); 
+            if (!AssetBundleModel.Model.DataSource.IsReadOnly())
+            {
+                menu.AddItem(new GUIContent("Add new bundle"), false, CreateNewBundle, selectedNodes);
                 menu.AddItem(new GUIContent("Add new folder"), false, CreateFolder, selectedNodes);
             }
             menu.AddSeparator(string.Empty);
 
-            menu.AddItem(new GUIContent("Export all data"), false, ExportABInfo, selectedNodes);
+            if (AssetDatabase.GetAllAssetBundleNames().Length != 0)
+            {
+                menu.AddItem(new GUIContent("Export all data"), false, ExportAllABInfo, selectedNodes);
+            }
+
             menu.AddItem(new GUIContent("Import data"), false, ImportABInfo, selectedNodes);
 
             menu.AddSeparator(string.Empty);
@@ -144,19 +149,49 @@ namespace AssetBundleBrowser
 
         private void ExportABInfo(object varContext)
         {
-
+            var tempBundleName = new List<string>();
+            var tmpSelectBundleInfo = varContext as List<AssetBundleModel.BundleTreeItem>;
+            tmpSelectBundleInfo.ForEach(bt => tempBundleName.Add(bt.bundle.displayName));
+            ExportBundleJson(tempBundleName);
+        }
+        private void ExportAllABInfo(object varContext)
+        {
+            ExportBundleJson(new List<string>(AssetDatabase.GetAllAssetBundleNames()));
         }
         private void ImportABInfo(object varContext)
         {
-            string tempOpenPath = EditorUtility.OpenFilePanel("Select AssetBuddle Info", Application.dataPath, "json");
-            
+            string tempOpenPath = EditorUtility.OpenFilePanel("Select AssetBuddle Info", Path.GetFullPath(Path.Combine(Application.dataPath, "../")), "json");
+
+            if (!File.Exists(tempOpenPath)) return;
+
+            var tempStr = File.ReadAllText(tempOpenPath);
+            var tmpAssetBundleInfo = JsonFx.Json.JsonReader.Deserialize<Dictionary<string, string>>(tempStr);
+
+            AssetDatabase.StartAssetEditing();
+            foreach (var tempkvp in tmpAssetBundleInfo)
+            {
+                var tempImporter = AssetImporter.GetAtPath(tempkvp.Key);
+                tempImporter.SetAssetBundleNameAndVariant(tempkvp.Value, string.Empty);
+            }
+            AssetDatabase.StopAssetEditing();
+
             ForceReloadData(null);
             AssetDatabase.RemoveUnusedAssetBundleNames();
         }
 
+        private void ExportBundleJson(List<string> varBundleNames)
+        {
+            var tempSavaPath = EditorUtility.SaveFilePanel("Sava Current AssetBuddle Info", Path.GetFullPath(Path.Combine(Application.dataPath, "../")), "ABInfo", "json");
+            if (MiscUtils.ExportBundleJson(tempSavaPath, varBundleNames))
+            {
+                EditorUtility.RevealInFinder(tempSavaPath);
+            }
+        }
+
         protected override void ContextClickedItem(int id)
         {
-            if (AssetBundleModel.Model.DataSource.IsReadOnly ()) {
+            if (AssetBundleModel.Model.DataSource.IsReadOnly())
+            {
                 return;
             }
 
@@ -166,10 +201,10 @@ namespace AssetBundleBrowser
             {
                 selectedNodes.Add(FindItem(nodeID, rootItem) as AssetBundleModel.BundleTreeItem);
             }
-            
+
             GenericMenu menu = new GenericMenu();
-            
-            if(selectedNodes.Count == 1)
+
+            if (selectedNodes.Count == 1)
             {
                 if ((selectedNodes[0].bundle as AssetBundleModel.BundleFolderConcreteInfo) != null)
                 {
@@ -178,7 +213,7 @@ namespace AssetBundleBrowser
                     menu.AddItem(new GUIContent("Add Sibling/New Bundle"), false, CreateNewSiblingBundle, selectedNodes);
                     menu.AddItem(new GUIContent("Add Sibling/New Folder"), false, CreateNewSiblingFolder, selectedNodes);
                 }
-                else if( (selectedNodes[0].bundle as AssetBundleModel.BundleVariantFolderInfo) != null)
+                else if ((selectedNodes[0].bundle as AssetBundleModel.BundleVariantFolderInfo) != null)
                 {
                     menu.AddItem(new GUIContent("Add Child/New Variant"), false, CreateNewVariant, selectedNodes);
                     menu.AddItem(new GUIContent("Add Sibling/New Bundle"), false, CreateNewSiblingBundle, selectedNodes);
@@ -198,18 +233,25 @@ namespace AssetBundleBrowser
                         menu.AddItem(new GUIContent("Add Sibling/New Variant"), false, CreateNewSiblingVariant, selectedNodes);
                     }
                 }
-                if(selectedNodes[0].bundle.IsMessageSet(MessageSystem.MessageFlag.AssetsDuplicatedInMultBundles))
+                if (selectedNodes[0].bundle.IsMessageSet(MessageSystem.MessageFlag.AssetsDuplicatedInMultBundles))
                     menu.AddItem(new GUIContent("Move duplicates to new bundle"), false, DedupeAllBundles, selectedNodes);
                 menu.AddItem(new GUIContent("Rename"), false, RenameBundle, selectedNodes);
                 menu.AddItem(new GUIContent("Delete " + selectedNodes[0].displayName), false, DeleteBundles, selectedNodes);
-                
+
             }
             else if (selectedNodes.Count > 1)
-            { 
+            {
                 menu.AddItem(new GUIContent("Move duplicates shared by selected"), false, DedupeOverlappedBundles, selectedNodes);
                 menu.AddItem(new GUIContent("Move duplicates existing in any selected"), false, DedupeAllBundles, selectedNodes);
                 menu.AddItem(new GUIContent("Delete " + selectedNodes.Count + " selected bundles"), false, DeleteBundles, selectedNodes);
             }
+
+            menu.AddSeparator(string.Empty);
+            menu.AddItem(new GUIContent("Import data"), false, ImportABInfo, selectedNodes);
+            menu.AddSeparator(string.Empty);
+            menu.AddItem(new GUIContent("Export Select Data"), false, ExportABInfo, selectedNodes);
+            menu.AddItem(new GUIContent("Export All Data"), false, ExportAllABInfo, selectedNodes);
+
             menu.ShowAsContext();
         }
         void ForceReloadData(object context)
@@ -340,7 +382,7 @@ namespace AssetBundleBrowser
         {
             var selectedNodes = context as List<AssetBundleModel.BundleTreeItem>;
             var newBundle = AssetBundleModel.Model.HandleDedupeBundles(selectedNodes.Select(item => item.bundle), onlyOverlappedAssets);
-            if(newBundle != null)
+            if (newBundle != null)
             {
                 var selection = new List<int>();
                 selection.Add(newBundle.nameHashCode);
@@ -412,7 +454,7 @@ namespace AssetBundleBrowser
                                 else
                                     hasNonScene = true;
 
-                                if ( (dataBundle as AssetBundleModel.BundleVariantDataInfo) != null)
+                                if ((dataBundle as AssetBundleModel.BundleVariantDataInfo) != null)
                                     hasVariantChild = true;
                             }
                         }
@@ -435,15 +477,16 @@ namespace AssetBundleBrowser
         {
             DragAndDropVisualMode visualMode = DragAndDropVisualMode.None;
             DragAndDropData data = new DragAndDropData(args);
-            
-            if (AssetBundleModel.Model.DataSource.IsReadOnly ()) {
+
+            if (AssetBundleModel.Model.DataSource.IsReadOnly())
+            {
                 return DragAndDropVisualMode.Rejected;
             }
 
-            if ( (data.hasScene && data.hasNonScene) ||
-                (data.hasVariantChild) )
+            if ((data.hasScene && data.hasNonScene) ||
+                (data.hasVariantChild))
                 return DragAndDropVisualMode.Rejected;
-            
+
             switch (args.dragAndDropPosition)
             {
                 case DragAndDropPosition.UponItem:
@@ -462,7 +505,7 @@ namespace AssetBundleBrowser
                             Reload();
                         }
                     }
-                    else if(data.paths != null)
+                    else if (data.paths != null)
                     {
                         visualMode = DragAndDropVisualMode.Copy;
                         if (data.args.performDrop)
@@ -483,7 +526,7 @@ namespace AssetBundleBrowser
             {
                 if (targetDataBundle.isSceneBundle)
                 {
-                    if(data.hasNonScene)
+                    if (data.hasNonScene)
                         return DragAndDropVisualMode.Rejected;
                 }
                 else
@@ -499,7 +542,7 @@ namespace AssetBundleBrowser
 
                 }
 
-               
+
                 if (data.args.performDrop)
                 {
                     if (data.draggedNodes != null)
@@ -536,7 +579,7 @@ namespace AssetBundleBrowser
                 }
                 else
                     visualMode = DragAndDropVisualMode.Rejected; //must be a variantfolder
-                
+
             }
             return visualMode;
         }
@@ -650,7 +693,7 @@ namespace AssetBundleBrowser
             var selection = new List<int>();
             selection.Add(hashCode);
             ReloadAndSelect(selection);
-            if(rename)
+            if (rename)
             {
                 BeginRename(FindItem(hashCode, rootItem), 0.25f);
             }
